move site/stage to package
