+++ core/src/test/java/org/dotwebstack/framework/site/SiteTest.java
@@ -1,10 +1,11 @@
-package org.dotwebstack.framework;
+package org.dotwebstack.framework.site;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org. junit.Assert.assertThat;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+import org.dotwebstack.framework.site.Site;
 import org.dotwebstack.framework.test.DBEERPEDIA;
 import org.junit.Test;
 import org.junit.runner.RunWith;
